[Gold IV] Title: 좋은 친구, Time: 244 ms, Memory: 4372 KB -BaekjoonHub

+++ 백준/Gold/3078. 좋은 친구/좋은 친구.cc
@@ -5,7 +5,7 @@ int main() {
     long long n, k, ans = 0;
     cin >> n >> k;
     vector<long long> a(n);
-    map<long long, long long> m;
+    map<long long, int> m;
     for (auto &i : a) {
         string s; cin >> s;
         i = s.length();
